refac: split write strategies into seperate modules
